+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 6.3.6
+### Feature:
+- Rewrite a part of the psql service to use new functionality.
+- Add `dns_name` to `ionoscloud_pg_cluster` datasource and resource
+### Dependency update
+- Update `go` to version 1.19
+- Update `sdk-go-dbaas-postgres` to [v1.0.6](https://github.com/ionos-cloud/sdk-go-dbaas-postgres/releases/tag/v1.0.6)
+### Documentation
+- Update documentation for K8s node pools
+
 ## 6.3.5
 ### Feature:
 - Removed EA note for container registry and dbaas mongo docs